moved BrainzPlayerDatabase to shared module
